# Conflicts:
#	trunk/source/AbstractInterpretationV2/src/de/uni_freiburg/informatik/ultimate/plugins/analysis/abstractinterpretationv2/algorithm/IDebugHelper.java
#	trunk/source/Library-ModelCheckerUtils/src/de/uni_freiburg/informatik/ultimate/lib/modelcheckerutils/absint/DisjunctiveAbstractState.java

+++ trunk/source/TraceAbstraction/src/de/uni_freiburg/informatik/ultimate/plugins/generator/traceabstraction/PathProgramCache.java
@@ -99,7 +99,7 @@ public class PathProgramCache<LETTER> {
 			mLogger.warn("You did not report this counterexample before!");
 			return 0;
 		}
-		return count.intValue();
+		return count;
 	}
 
 	public List<Integer> computeSortedHistrogram() {
